The declaration of following yarp classes have been moved from libYARP_math to libYARP_sig. Quaternion FrameTransform Vec2D
The implementation of methods which do not use libYARP_math functionalities are defined in libYARP_sig, libYARP_math otherwise.
