+++ src/libYARP_math/src/Quaternion.cpp
@@ -6,151 +6,14 @@
  * BSD-3-Clause license. See the accompanying LICENSE file for details.
  */
 
-#include <yarp/math/Quaternion.h>
+#include <yarp/sig/Quaternion.h>
 #include <yarp/math/Math.h>
 #include <cmath>
 #include <cstdio>
 
+using namespace yarp::sig;
 using namespace yarp::math;
-
-YARP_BEGIN_PACK
-class QuaternionPortContentHeader
-{
-public:
-    yarp::os::NetInt32 listTag;
-    yarp::os::NetInt32 listLen;
-    QuaternionPortContentHeader() : listTag(0), listLen(0) {}
-};
-YARP_END_PACK
-
-Quaternion::Quaternion()
-{
-    internal_data[0] = 1;
-    internal_data[1] = 0;
-    internal_data[2] = 0;
-    internal_data[3] = 0;
-}
-
-Quaternion::Quaternion(double x, double y, double z, double w)
-{
-    internal_data[0] = w;
-    internal_data[1] = x;
-    internal_data[2] = y;
-    internal_data[3] = z;
-}
-
-Quaternion::Quaternion(const Quaternion &l)
-{
-    internal_data[0] = l.w();
-    internal_data[1] = l.x();
-    internal_data[2] = l.y();
-    internal_data[3] = l.z();
-}
-
-const double* Quaternion::data() const
-{
-    return internal_data;
-}
-
-double* Quaternion::data()
-{
-    return internal_data;
-}
-
-yarp::sig::Vector Quaternion::toVector()  const
-{
-    yarp::sig::Vector v(4);
-    v[0] = internal_data[0];
-    v[1] = internal_data[1];
-    v[2] = internal_data[2];
-    v[3] = internal_data[3];
-    return v;
-}
-
-double Quaternion::w() const
-{
-    return internal_data[0];
-}
-
-double Quaternion::x() const
-{
-    return internal_data[1];
-}
-
-double Quaternion::y() const
-{
-    return internal_data[2];
-}
-
-double Quaternion::z() const
-{
-    return internal_data[3];
-}
-
-double& Quaternion::w()
-{
-    return internal_data[0];
-}
-
-double& Quaternion::x()
-{
-    return internal_data[1];
-}
-
-double& Quaternion::y()
-{
-    return internal_data[2];
-}
-
-double& Quaternion::z()
-{
-    return internal_data[3];
-}
-
-bool Quaternion::read(yarp::os::ConnectionReader& connection)
-{
-    // auto-convert text mode interaction
-    connection.convertTextMode();
-    QuaternionPortContentHeader header;
-    bool ok = connection.expectBlock((char*)&header, sizeof(header));
-    if (!ok) return false;
-
-    if (header.listLen == 4 &&  header.listTag == (BOTTLE_TAG_LIST | BOTTLE_TAG_FLOAT64))
-    {
-        this->internal_data[0] = connection.expectFloat64();
-        this->internal_data[1] = connection.expectFloat64();
-        this->internal_data[2] = connection.expectFloat64();
-        this->internal_data[3] = connection.expectFloat64();
-    }
-    else
-    {
-        return false;
-    }
-
-    return !connection.isError();
-}
-
-bool Quaternion::write(yarp::os::ConnectionWriter& connection)
-{
-    QuaternionPortContentHeader header;
-
-    header.listTag = (BOTTLE_TAG_LIST | BOTTLE_TAG_FLOAT64);
-    header.listLen = 4;
-
-    connection.appendBlock((char*)&header, sizeof(header));
-
-    connection.appendFloat64(this->internal_data[0]);
-    connection.appendFloat64(this->internal_data[1]);
-    connection.appendFloat64(this->internal_data[2]);
-    connection.appendFloat64(this->internal_data[3]);
-
-    // if someone is foolish enough to connect in text mode,
-    // let them see something readable.
-    connection.convertTextMode();
-
-    return !connection.isError();
-}
-
+/*
 void Quaternion::fromRotationMatrix(const yarp::sig::Matrix &R)
 {
     if ((R.rows()<3) || (R.cols()<3))
@@ -206,7 +69,7 @@ void Quaternion::fromRotationMatrix(const yarp::sig::Matrix &R)
         internal_data[2] = (R(1, 0) + R(0, 1))*sqdip1;
         internal_data[3] = (R(0, 2) + R(2, 0))*sqdip1;
     }
-}
+}*/
 
 yarp::sig::Matrix Quaternion::toRotationMatrix4x4() const
 {
@@ -299,34 +162,3 @@ yarp::sig::Vector Quaternion::toAxisAngle()
     return v;
 }
 
-double Quaternion::abs()
-{
-    return sqrt(internal_data[0] * internal_data[0] +
-                internal_data[1] * internal_data[1] +
-                internal_data[2] * internal_data[2] +
-                internal_data[3] * internal_data[3]);
-}
-
-void Quaternion::normalize()
-{
-    double length = abs();
-    internal_data[0] /= length;
-    internal_data[1] /= length;
-    internal_data[2] /= length;
-    internal_data[3] /= length;
-    return;
-}
-
-double Quaternion::arg()
-{
-    return atan2(sqrt(internal_data[1] * internal_data[1] +
-                      internal_data[2] * internal_data[2] +
-                      internal_data[3] * internal_data[3]),
-                 internal_data[0]);
-}
-
-Quaternion Quaternion::inverse() const
-{
-    //                     w                  x                 y                  z
-    return Quaternion(internal_data[0], -internal_data[1], -internal_data[2], -internal_data[3]);
-}
