+++ src/libYARP_math/src/Vec2D.cpp
@@ -6,7 +6,7 @@
  * BSD-3-Clause license. See the accompanying LICENSE file for details.
  */
 
-#include <yarp/math/Vec2D.h>
+#include <yarp/sig/Vec2D.h>
 #include <yarp/math/Math.h>
 #include <sstream>
 #include <cmath>
@@ -15,143 +15,12 @@
 // network stuff
 #include <yarp/os/NetInt32.h>
 
+using namespace yarp::sig;
 using namespace yarp::math;
 
-YARP_BEGIN_PACK
-class Vec2DPortContentHeader
-{
-public:
-    yarp::os::NetInt32 listTag;
-    yarp::os::NetInt32 listLen;
-    Vec2DPortContentHeader() : listTag(0), listLen(0) {}
-};
-YARP_END_PACK
-
-namespace yarp {
-namespace math {
-template<>
-bool Vec2D<double>::read(yarp::os::ConnectionReader& connection)
-{
-    // auto-convert text mode interaction
-    connection.convertTextMode();
-    Vec2DPortContentHeader header;
-    bool ok = connection.expectBlock((char*)&header, sizeof(header));
-    if (!ok) return false;
-
-    if (header.listLen == 2 && header.listTag == (BOTTLE_TAG_LIST | BOTTLE_TAG_FLOAT64))
-    {
-        this->x = connection.expectFloat64();
-        this->y = connection.expectFloat64();
-    }
-    else
-    {
-        return false;
-    }
-
-    return !connection.isError();
-}
-
-template<>
-bool Vec2D<int>::read(yarp::os::ConnectionReader& connection)
-{
-    // auto-convert text mode interaction
-    connection.convertTextMode();
-    Vec2DPortContentHeader header;
-    bool ok = connection.expectBlock((char*)&header, sizeof(header));
-    if (!ok) return false;
-
-    if (header.listLen == 2 && header.listTag == (BOTTLE_TAG_LIST | BOTTLE_TAG_INT32))
-    {
-        this->x = connection.expectInt32();
-        this->y = connection.expectInt32();
-    }
-    else
-    {
-        return false;
-    }
-
-    return !connection.isError();
-}
-
-template<>
-bool Vec2D<double>::write(yarp::os::ConnectionWriter& connection)
-{
-    Vec2DPortContentHeader header;
-
-    header.listTag = (BOTTLE_TAG_LIST | BOTTLE_TAG_FLOAT64);
-    header.listLen = 2;
-
-    connection.appendBlock((char*)&header, sizeof(header));
-
-    connection.appendFloat64(this->x);
-    connection.appendFloat64(this->y);
-
-    connection.convertTextMode();
-
-    return !connection.isError();
-}
-
-template<>
-bool Vec2D<int>::write(yarp::os::ConnectionWriter& connection)
-{
-    Vec2DPortContentHeader header;
-
-    header.listTag = (BOTTLE_TAG_LIST | BOTTLE_TAG_INT32);
-    header.listLen = 2;
-
-    connection.appendBlock((char*)&header, sizeof(header));
-
-    connection.appendInt32(this->x);
-    connection.appendInt32(this->y);
-
-    connection.convertTextMode();
-
-    return !connection.isError();
-}
-
-} // namespace math
-} // namespace yarp
-
-
+/*
 template <typename T>
-std::string yarp::math::Vec2D<T>::toString(int precision, int width) const
-{
-    std::ostringstream stringStream;
-    stringStream.precision(precision);
-    stringStream.width(width);
-    stringStream << std::string("x:") << x << std::string(" y:") << y;
-    return stringStream.str();
-}
-
-template <typename T>
-T yarp::math::Vec2D<T>::norm() const
-{
-    return T(sqrt(x*x + y*y));
-}
-
-//constructors
-template <typename T>
-yarp::math::Vec2D<T>::Vec2D() : x(0), y(0)
-{
-}
-
-template <typename T>
-yarp::math::Vec2D<T>::Vec2D(const yarp::sig::Vector& v)
-{
-    yAssert(v.size() == 2);
-    x = T(v[0]);
-    y = T(v[1]);
-}
-
-template <typename T>
-yarp::math::Vec2D<T>::Vec2D(const T& x_value, const T& y_value)
-{
-    x = x_value;
-    y = y_value;
-}
-
-template <typename T>
- yarp::math::Vec2D<T>  operator * (const yarp::sig::Matrix& lhs, yarp::math::Vec2D<T> rhs)
+ yarp::sig::Vec2D<T>  operator * (const yarp::sig::Matrix& lhs, yarp::sig::Vec2D<T> rhs)
 {
     yAssert(lhs.rows() == 2 && lhs.cols() == 2);
     T x = rhs.x; T y = rhs.y;
@@ -161,21 +30,21 @@ template <typename T>
 }
 
 template <typename T>
-yarp::math::Vec2D<T> operator + (yarp::math::Vec2D<T> lhs, const yarp::math::Vec2D<T>& rhs)
+yarp::sig::Vec2D<T> operator + (yarp::sig::Vec2D<T> lhs, const yarp::sig::Vec2D<T>& rhs)
 {
     lhs += rhs;
     return lhs;
 }
 
 template <typename T>
-yarp::math::Vec2D<T> operator - (yarp::math::Vec2D<T> lhs, const yarp::math::Vec2D<T>& rhs)
+yarp::sig::Vec2D<T> operator - (yarp::sig::Vec2D<T> lhs, const yarp::sig::Vec2D<T>& rhs)
 {
     lhs -= rhs;
     return lhs;
 }
 
 template <typename T>
-yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator+=(const yarp::math::Vec2D<T>& rhs)
+yarp::sig::Vec2D<T>& yarp::sig::Vec2D<T>::operator+=(const yarp::sig::Vec2D<T>& rhs)
 {
     this->x += rhs.x;
     this->y += rhs.y;
@@ -183,7 +52,7 @@ yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator+=(const yarp::math::Vec2D<T
 }
 
 template <typename T>
-yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator-=(const yarp::math::Vec2D<T>& rhs)
+yarp::sig::Vec2D<T>& yarp::sig::Vec2D<T>::operator-=(const yarp::sig::Vec2D<T>& rhs)
 {
     this->x -= rhs.x;
     this->y -= rhs.y;
@@ -191,7 +60,7 @@ yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator-=(const yarp::math::Vec2D<T
 }
 
 template <typename T>
-yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator =(const yarp::math::Vec2D<T>& rhs)
+yarp::sig::Vec2D<T>& yarp::sig::Vec2D<T>::operator =(const yarp::sig::Vec2D<T>& rhs)
 {
     if (this != &rhs)
     {
@@ -202,7 +71,7 @@ yarp::math::Vec2D<T>& yarp::math::Vec2D<T>::operator =(const yarp::math::Vec2D<T
 }
 
 template <typename T>
-bool yarp::math::Vec2D<T>::operator ==(const yarp::math::Vec2D<T>& rhs)
+bool yarp::sig::Vec2D<T>::operator ==(const yarp::sig::Vec2D<T>& rhs)
 {
     if (this->x == rhs.x &&
         this->y == rhs.y) return true;
@@ -210,19 +79,20 @@ bool yarp::math::Vec2D<T>::operator ==(const yarp::math::Vec2D<T>& rhs)
 }
 
 template <typename T>
-bool yarp::math::Vec2D<T>::operator !=(const yarp::math::Vec2D<T>& rhs)
+bool yarp::sig::Vec2D<T>::operator !=(const yarp::sig::Vec2D<T>& rhs)
 {
     if (this->x == rhs.x &&
         this->y == rhs.y) return false;
     return true;
 }
 
-template yarp::math::Vec2D<double> YARP_math_API operator + (yarp::math::Vec2D<double> lhs, const yarp::math::Vec2D<double>& rhs);
-template yarp::math::Vec2D<int>    YARP_math_API operator + (yarp::math::Vec2D<int> lhs, const yarp::math::Vec2D<int>& rhs);
-template yarp::math::Vec2D<double> YARP_math_API operator - (yarp::math::Vec2D<double> lhs, const yarp::math::Vec2D<double>& rhs);
-template yarp::math::Vec2D<int>    YARP_math_API operator - (yarp::math::Vec2D<int> lhs, const yarp::math::Vec2D<int>& rhs);
-template yarp::math::Vec2D<double> YARP_math_API operator * (const yarp::sig::Matrix& lhs, yarp::math::Vec2D<double> rhs);
-template yarp::math::Vec2D<int>    YARP_math_API operator * (const yarp::sig::Matrix& lhs, yarp::math::Vec2D<int> rhs);
+template yarp::sig::Vec2D<double> YARP_sig_API operator + (yarp::sig::Vec2D<double> lhs, const yarp::sig::Vec2D<double>& rhs);
+template yarp::sig::Vec2D<int>    YARP_sig_API operator + (yarp::sig::Vec2D<int> lhs, const yarp::sig::Vec2D<int>& rhs);
+template yarp::sig::Vec2D<double> YARP_sig_API operator - (yarp::sig::Vec2D<double> lhs, const yarp::sig::Vec2D<double>& rhs);
+template yarp::sig::Vec2D<int>    YARP_sig_API operator - (yarp::sig::Vec2D<int> lhs, const yarp::sig::Vec2D<int>& rhs);
+template yarp::sig::Vec2D<double> YARP_sig_API operator * (const yarp::sig::Matrix& lhs, yarp::sig::Vec2D<double> rhs);
+template yarp::sig::Vec2D<int>    YARP_sig_API operator * (const yarp::sig::Matrix& lhs, yarp::sig::Vec2D<int> rhs);
 
-template class YARP_math_API yarp::math::Vec2D<double>;
-template class YARP_math_API yarp::math::Vec2D<int>;
+template class YARP_sig_API yarp::sig::Vec2D<double>;
+template class YARP_sig_API yarp::sig::Vec2D<int>;
+*/
